Fix Legend Clipping into Charts
When a chart with a Legend that was positioned above it was rendered
initially, the calculations for the height of the chart itself would not
take into consideration the height of the container for the Legend,
resulting in a height that did not have the correct offset.

This appears to be caused by `setState` calls being batched together as
React is attempting to gain performance wins.

Previously, the fix for this was to call `requestAnimationFrame` when we
set the internal state on `componentDidMount` to ensure it is called as
soon as possible, rather than being batched in with other calls.

However, there is now a PR open in the main Recharts repository that
handles the change in a slightly different, but cleaner manner. This PR
is not updated to reflect that.

PR in question: https://github.com/recharts/recharts/pull/2062